+++ src/Codeception/Command/SelfUpdate.php
@@ -5,6 +5,7 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
+use Codeception\Codecept;
 
 /**
  * Auto-updates phar archive from official site: 'http://codeception.com/codecept.phar' .
@@ -19,9 +20,9 @@ class SelfUpdate extends Command
      * Class constants
      */
     const NAME = 'Codeception';
-    const GITHUB = 'Codeception/Codeception';
-    const SOURCE = 'http://codeception.com/codecept.phar';
-    const SOURCE_PHP54 = 'http://codeception.com/php54/codecept.phar';
+    const GITHUB_REPO = 'Codeception/Codeception';
+    const PHAR_URL = 'http://codeception.com/releases/%s/codecept.phar';
+    const PHAR_URL_PHP54 = 'http://codeception.com/releases/%s/php54/codecept.phar';
 
     /**
      * Holds the current script filename.
@@ -55,26 +56,36 @@ class SelfUpdate extends Command
     }
 
     /**
+     * @return string
+     */
+    protected function getCurrentVersion()
+    {
+        return Codecept::VERSION;
+    }
+
+    /**
      * {@inheritdoc}
      */
     public function execute(InputInterface $input, OutputInterface $output)
     {
-        $version = \Codeception\Codecept::VERSION;
+        $version = $this->getCurrentVersion();
 
         $output->writeln(
             sprintf(
                 '<info>%s</info> version <comment>%s</comment>',
-                self::NAME, $version
+                self::NAME,
+                $version
             )
         );
 
         $output->writeln("\n<info>Checking for a new version...</info>\n");
         try {
-            if ($this->isOutOfDate($version)) {
+            $latestVersion = $this->getLatestStableVersion();
+            if ($this->isOutOfDate($version, $latestVersion)) {
                 $output->writeln(
                     sprintf(
                         'A newer version is available: <comment>%s</comment>',
-                        $this->liveVersion
+                        $latestVersion
                     )
                 );
                 if (!$input->getOption('no-interaction')) {
@@ -89,7 +100,7 @@ class SelfUpdate extends Command
                 }
                 $output->writeln("\n<info>Updating...</info>");
 
-                $this->retrieveLatestPharFile($output);
+                $this->retrievePharFile($latestVersion, $output);
 
             } else {
                 $output->writeln('You are already using the latest version.');
@@ -109,21 +120,41 @@ class SelfUpdate extends Command
     /**
      * Checks whether the provided version is current.
      *
-     * @param  string $version The version number to check.
+     * @param string $version The version number to check.
+     * @param string $latestVersion Latest stable version
      * @return boolean Returns True if a new version is available.
      */
-    private function isOutOfDate($version)
+    private function isOutOfDate($version, $latestVersion)
     {
-        $tags = $this->getGithubTags(self::GITHUB);
+        return -1 != version_compare($version, $latestVersion, '>=');
+    }
 
-        $this->liveVersion = array_reduce(
-            $tags,
+    /**
+     * @return string
+     */
+    private function getLatestStableVersion()
+    {
+        $stableVersions = $this->filterStableVersions(
+            $this->getGithubTags(self::GITHUB_REPO)
+        );
+
+        return array_reduce(
+            $stableVersions,
             function ($a, $b) {
                 return version_compare($a, $b, '>') ? $a : $b;
             }
         );
+    }
 
-        return -1 != version_compare($version, $this->liveVersion, '>=');
+    /**
+     * @param array $tags
+     * @return array
+     */
+    private function filterStableVersions($tags)
+    {
+        return array_filter($tags, function ($tag) {
+            return preg_match('/^[0-9]+\.[0-9]+\.[0-9]+$/', $tag);
+        });
     }
 
     /**
@@ -132,7 +163,7 @@ class SelfUpdate extends Command
      * @param  string $repo The repository name to check upon.
      * @return array
      */
-    private function getGithubTags($repo)
+    protected function getGithubTags($repo)
     {
         $jsonTags = $this->retrieveContentFromUrl(
             'https://api.github.com/repos/' . $repo . '/tags'
@@ -151,7 +182,7 @@ class SelfUpdate extends Command
      *
      * @param  string $url
      * @return string
-     * @throw Exception if status code is above 300
+     * @throws \Exception if status code is above 300
      */
     private function retrieveContentFromUrl($url)
     {
@@ -218,19 +249,17 @@ class SelfUpdate extends Command
     /**
      * Retrieves the latest phar file.
      *
+     * @param string $version
      * @param OutputInterface $output
+     * @throws \Exception
      */
-    protected function retrieveLatestPharFile(OutputInterface $output)
+    protected function retrievePharFile($version, OutputInterface $output)
     {
         $temp = basename($this->filename, '.phar') . '-temp.phar';
 
         try {
-            $source = self::SOURCE;
-            if (version_compare(PHP_VERSION, '5.6.0', '<')) {
-                $source = self::SOURCE_PHP54;
-            }
-
-            if (@copy($source, $temp)) {
+            $sourceUrl = $this->getPharUrl($version);
+            if (@copy($sourceUrl, $temp)) {
                 chmod($temp, 0777 & ~umask());
 
                 // test the phar validity
@@ -242,8 +271,7 @@ class SelfUpdate extends Command
                 throw new \Exception('Request failed.');
             }
         } catch (\Exception $e) {
-            if (
-                !$e instanceof \UnexpectedValueException
+            if (!$e instanceof \UnexpectedValueException
                 && !$e instanceof \PharException
             ) {
                 throw $e;
@@ -266,4 +294,19 @@ class SelfUpdate extends Command
         );
     }
 
+    /**
+     * Returns Phar file URL for specified version
+     *
+     * @param string $version
+     * @return string
+     */
+    protected function getPharUrl($version)
+    {
+        $sourceUrl = self::PHAR_URL;
+        if (version_compare(PHP_VERSION, '5.6.0', '<')) {
+            $sourceUrl = self::PHAR_URL_PHP54;
+        }
+
+        return sprintf($sourceUrl, $version);
+    }
 }